refactor: update SVG imports to use new syntax for React components
