+++ src/components/Layout/Header.tsx
@@ -5,9 +5,9 @@ import { CgTab } from 'react-icons/cg'
 import { IoMdSunny } from 'react-icons/io'
 import { MdDoDisturbOff } from 'react-icons/md'
 import { Link, useLocation, useNavigate } from 'react-router-dom'
-import { ReactComponent as AvatarPlaceholder } from 'src/assets/icons/avatar.svg'
-import { ReactComponent as StreakIcon } from 'src/assets/icons/fire_icon.svg'
-import { ReactComponent as HackertabLogo } from 'src/assets/logo.svg'
+import AvatarPlaceholder from 'src/assets/icons/avatar.svg?react'
+import StreakIcon from 'src/assets/icons/fire_icon.svg?react'
+import HackertabLogo from 'src/assets/logo.svg?react'
 import { SearchBar } from 'src/components/Elements/SearchBar'
 import { UserTags } from 'src/components/Elements/UserTags'
 import { useAuth } from 'src/features/auth'
