+++ src/features/settings/components/UserSettings/UserInfo.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react'
 import { FaGithub } from 'react-icons/fa'
 import { FcGoogle } from 'react-icons/fc'
 import { IoCheckmarkOutline } from 'react-icons/io5'
-import { ReactComponent as AvatarPlaceholder } from 'src/assets/icons/avatar.svg'
-import { ReactComponent as StreakIcon } from 'src/assets/icons/fire_icon.svg'
+import AvatarPlaceholder from 'src/assets/icons/avatar.svg?react'
+import StreakIcon from 'src/assets/icons/fire_icon.svg?react'
 import { Button, ConfirmModal } from 'src/components/Elements'
 import { useAuth } from 'src/features/auth'
 import { pluralize } from 'src/utils/String'
